inspired runner lite

+++ data_collection/src/inspired_exerciser.py
@@ -8,8 +8,9 @@ if __name__ == '__main__':
     ISOTIMEFORMAT = '%m%d-%H-%M-%S'
     logger = Utilities.set_logger('COSMOS_TRIGGER_PY-Console')
 
-    device = 'nexus4'
+    device = 'xiaoyao'
     pc = 'iai'
+    lite = True
 
     if device == 'nexus4':
         series = '01b7006e13dd12a1'
@@ -17,13 +18,15 @@ if __name__ == '__main__':
         series = '014E233C1300800B'
     elif device == 'nexuss':
         series = '39302E8CEA9B00EC'
+    elif device == 'xiaoyao':
+        series = '127.0.0.1:21503'
     else:
-        series = 'emulator-5554'
+        series = 'emulator-5556'
 
 
-    user = 'hfu'
-    aapt_loc = 'C:\Users\\' + user + '\AppData\Local\Android\sdk/build-tools/19.1.0/aapt.exe'
-    apk_dir = 'C:\Users\\' + user + '\Documents\FlowIntent\\apks\\Books_Reference\\'
+    user = 'Hao'
+    aapt_loc = 'C:\Users\\' + user + '\AppData\Local\Android\sdk/build-tools/25.0.0/aapt.exe'
+    apk_dir = 'C:\Users\\' + user + '\Documents\COSMOS\\apk\\Tools\\'
     UIExerciser.emu_loc = 'C:\Users\hfu\AppData\Local\Android\sdk/tools/emulator.exe'
     UIExerciser.emu_name = 'Qvga'
 
@@ -39,11 +42,20 @@ if __name__ == '__main__':
                 # main_process.join()
                 for i in range(3):
                     try:
+                        if device == 'xiaoyao':
+                            UIExerciser.run_adb_cmd('connect 127.0.0.1:21503')
                         apk = os.path.join(root, filename)
                         exerciser = UIExerciser(series, aapt_loc, apk_dir, out_base_dir, logger)
-                        exerciser.inspired_run(series, os.path.join(root, filename), examined, 'C:\\Users\\hfu\\Documents\\COSMOS\\output\\java\\')
+                        if lite:
+                            exerciser.inspired_run_lite(series, os.path.join(root, filename), examined, 'C:\\Users\\Hao\\Documents\\COSMOS\\output\\java\\')
+                        else:
+                            exerciser.inspired_run(series, os.path.join(root, filename), examined,
+                                                   'C:\\Users\\hfu\\Documents\\COSMOS\\output\\java\\')
                         break
                     except Exception as e:
                         print e
-                        UIExerciser.run_adb_cmd('reboot')
+                        if device == 'xiaoyao':
+                            UIExerciser.run_adb_cmd('connect 127.0.0.1:21503')
+                        else:
+                            UIExerciser.run_adb_cmd('reboot')
                         time.sleep(90)
